# internal/apiserver/controller/v1/user/password.go
package user

import (
	"github.com/asaskevich/govalidator"
	"github.com/gin-gonic/gin"

	"demo/internal/pkg/core"
	"demo/internal/pkg/errno"
	"demo/internal/pkg/log"
	v1 "demo/pkg/api/demo/v1"
)

// ChangePassword 修改指定用户的密码.
//
// @Summary    Change password
// @Security   Bearer
// @Tags       User
// @Accept     application/json
// @Produce    json
// @Param      name	     path	    string          	        true  "Username"
// @Param      request	 body	    v1.ChangePasswordRequest	true  "Param"
// @Success	   200		{object}	nil
// @Failure	   400		{object}	core.ErrResponse
// @Failure	   500		{object}	core.ErrResponse
// @Router    /v1/users/{name}/change-password [PUT].
func (ctrl *UserController) ChangePassword(c *gin.Context) {
	log.C(c).Infow("Change password function called")

	var r v1.ChangePasswordRequest
	if err := c.ShouldBindJSON(&r); err != nil {
		core.WriteResponse(c, errno.ErrBind, nil)

		return
	}

	if _, err := govalidator.ValidateStruct(r); err != nil {
		core.WriteResponse(c, errno.ErrInvalidParameter.SetMessage(err.Error()), nil)

		return
	}

	username := c.Param("name")
	err := ctrl.b.Users().ChangePassword(c, username, &r)
	if err != nil {
		core.WriteResponse(c, err, nil)

		return
	}

	core.WriteResponse(c, nil, nil)
}